LayoutTests/fast/harness/memoryinfo-object-expected.txt

PASS [object Internals] is defined.
PASS [object MemoryInfo] is defined.
PASS stats.usedJSHeapSize is >= 0
PASS stats.totalJSHeapSize is >= 0
